IonBase

Foundation of IonCLI.

References

IonBase.PRN — Type

PRN{name}

Package Registry Name

IonBase.PRN — Method

PRN(name::String)

Create a PRN (Pacakge Registry Name) object.

IonBase.add — Function

add package/project to the closest project.

Arguments

urls: package names or urls to add.

Flags

-g, --glob: add package to global shared environment.

IonBase.build — Function

build package/project/environment

Arguments

pkg: package name.

Flags

-v, --verbose: print verbose log

-g, --glob: enable to build in global shared environment

IonBase.clone — Function

clone a package or url and setup the local repo. If the current local git user do not have push access to remote github repo, it will fork this repo then clone the repo and set the input url as upstream.

Arguments

url_or_package: name of the package or url.

IonBase.create — Function

create a project or package.

Arguments

path: path of the project you want to create

Options

--user <name>: your GitHub user name for this package.

--template <template name>: template name.

Flags

-i, --interactive: enable to start interactive configuration interface.

IonBase.dev — Function

Make a package available for development. If pkg is an existing local path, that path will be recorded in the manifest and used. Otherwise, a full git clone of pkg is made. Unlike the dev/develop command in Julia REPL pkg mode, ion will clone the package to the dev folder of the current project. You can specify --shared flag to use shared dev folder under ~/.julia/dev (specified by JULIA_PKG_DEVDIR).

Arguments

url: URL or local path to the package.

Flags

-s, --shared: controls whether to use the shared develop folder.

IonBase.free — Function

If pkg is pinned, remove the pin. If pkg is tracking a path, e.g. after Pkg.develop, go back to tracking registered versions.

Flags

-g, --glob: enable to execute in global shared environment

IonBase.install — Function

install a given target. Currently only supports "path", future will support installing a specific Julia version and Comonicon applications.

Arguments

target: install given target, can be "path".

IonBase.instantiate — Function

If a Manifest.toml file exists in the active project, download all the packages declared in that manifest. Otherwise, resolve a set of feasible packages from the Project.toml files and install them. If no Project.toml exist in the current active project, create one with all the dependencies in the manifest and instantiate the resulting project.

Flags

-v, --verbose: prints the build output to stdout/stderr instead of redirecting to the build.log file.

IonBase.isurl — Method

isurl(str)

Checks if the given string is an absolute URL.

Examples

julia> isurl("https://julialang.org")
true
julia> isurl("mailto:someone@example.com")
true
julia> isurl("/foo/bar")
false

IonBase.redo — Function

Redoes the changes from the latest undo.

Flags

-g, --glob: enable to execute in global shared environment

IonBase.release — Function

release a package.

Arguments

version: version number you want to release. Can be a specific version, "current" or either of (major, minor, patch)

path: path to the project you want to release.

Options

-r,--registry <registry name>: registry you want to register the package. If the package has not been registered, ion will try to register the package in the General registry. Or the user needs to specify the registry to register using this option.

-b, --branch <branch name>: branch you want to register.

Flags

-q,--quiet: do not promote anything.

IonBase.resolve — Function

Update the current manifest. It will update manifest with potential changes to the dependency graph from packages that are tracking a path.

Flags

-g, --glob: enable to resolve in global shared environment

IonBase.rm — Function

Remove a package from the current project. If the mode of pkg is PKGMODE_MANIFEST also remove it from the manifest including all recursive dependencies of pkg.

Arguments

pkg: package name to remove.

Flags

-g, --glob: enable to remove package in global shared environment

IonBase.search — Function

search a package.

Options

--registry <name>: the name of registry you want to search, by default will search all registry in default DEPOT_PATH

--token <github_token>: specify GitHub authenticate token, this is necessary when you want to search frequently.

Flags

--omit: use this flag to print only the package names.

IonBase.status — Function

show current environment status

Arguments

pkg: package name.

Flags

-d, --diff: show diff to last git commit

-g, --glob: enable to show status in global shared environment

IonBase.test — Function

test package/project

Arguments

pkg: package name.

Flags

-g, --glob: enable to test in global shared environment

IonBase.undo — Function

Undoes the latest change to the active project. Only states in the current session are stored, up to a maximum of 50 states.

Flags

-g, --glob: enable to execute in global shared environment

IonBase.update — Function

Update a package. If no posistional argument is given, update all packages in current project.

Arguments

pkg: package name.

IonBase.Registry — Module

registry tools

IonBase.Registry.add — Function

add a registry

Arguments

url: URL to the registry, or use name "General" to add the default general registry.

IonBase.Doc — Module

documentation tools.

IonBase.Doc.build — Function

build documentation.

Args

path: path of the project.

IonBase.Doc.serve — Function

serve documentation.

Options

--foldername <name>: specify the name of the content folder if different than "docs".

--literate <path>: is the path to the folder containing the literate scripts, if left empty, it will be assumed that they are in docs/src.

Flags

--verbose: show verbose log.

--doc-env: is a boolean switch to make the server start by activating the doc environment or not (i.e. the Project.toml in docs/).